+++ src/txxt/parsers/reference.rs
@@ -5,14 +5,16 @@
 //!
 //! ## Structure
 //!
+//! - `builders.rs`: Consolidated AST node building functions and location utilities
 //! - `parser.rs`: Main parser orchestration
-//! - `combinators.rs`: Parser combinators and primitives
+//! - `combinators.rs`: Parser combinators and primitives (legacy, to be deprecated)
 //! - `api.rs`: Public API entry points
 //! - Element parsers (annotations, definitions, document, etc.): Build element-specific parsers
 //!   and construct AST nodes for each txxt element type
 
 pub mod annotations;
 pub mod api;
+pub mod builders;
 pub mod combinators;
 pub mod definitions;
 pub mod document;
